restructure: organize ui/ by elements, layout and page sections
