+++ app/blog/page.tsx
@@ -1,8 +1,8 @@
 import { type Metadata } from 'next'
 import { type ReactElement } from 'react'
 
-import PageLayout from '@/ui/layouts/page-layout'
-import PostList from '@/ui/post-list'
+import PageLayout from '@/ui/layout/main'
+import PostList from '@/ui/sections/blog-post-list'
 import { DEFAULT_OG_IMAGE, SITE_NAME, SITE_URL, TWITTER_HANDLE, SITE_LOCALE } from '@/seo/constants'
 import JsonLdScript from '@/seo/json-ld/script'
 
@@ -29,7 +29,10 @@ export default async function Blog(): Promise<ReactElement> {
   return (
     <PageLayout>
       <h1 className="sr-only">Blog</h1>
-      <PostList />
+      <section>
+        <h2 className="sr-only">Blog posts</h2>
+        <PostList />
+      </section>
       <JsonLdScript type="blog" />
     </PageLayout>
   )
