+++ app/not-found.test.tsx
@@ -7,7 +7,7 @@ import NotFound from './not-found'
 
 // Mock PageLayout to avoid async data fetching in Header/Footer
 // But preserve the main wrapper that PageLayout now provides
-vi.mock('@/ui/layouts/page-layout', () => ({
+vi.mock('@/ui/layout/main', () => ({
   default: ({ children }: { children: React.ReactNode }) => (
     <main id="main" className="flex-auto flex flex-col">
       {children}
