+++ app/page.tsx
@@ -1,7 +1,7 @@
 import { type ReactElement } from 'react'
-import PageLayout from '@/ui/layouts/page-layout'
-import Summary from '@/ui/home/summary'
-import RecentWriting from '@/ui/home/recent-writing'
+import PageLayout from '@/ui/layout/main'
+import Summary from '@/ui/sections/home-summary'
+import RecentWriting from '@/ui/sections/home-recent-writing'
 import JsonLdScript from '@/seo/json-ld/script'
 
 export default async function Home(): Promise<ReactElement> {
