+++ ui/sections/blog-post-list.tsx
@@ -1,8 +1,8 @@
 import { type ReactElement } from 'react'
 
 import getPosts from '@/io/notion/getPosts'
-import Link from '@/ui/link'
-import { getHumanReadableDate, getMachineReadableDate } from '@/ui/dates'
+import Link from '@/ui/elements/link'
+import { getHumanReadableDate, getMachineReadableDate } from '@/utils/dates'
 
 type PostListProps = Readonly<{
   limit?: number
